Added a fully working volume slider
Originally anything more than zero would have switched the volume to 1,
which is max. Now added a minimum and a maximum of 0 and 1 respectively,
and a smallchange property of 0.1.

+++ WpfMediaDB/audioPlayer.xaml.cs
@@ -130,7 +130,7 @@ namespace WpfMediaDB
         private void volumeSlider_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
         {
             // this single line means that the volume of the music will be determined by the set values of the slider
-            aPlayer.Volume = volumeSlider.Value;
+            aPlayer.Volume = (double)volumeSlider.Value;
         }
     }
 }
